requirements.txt
langchain
langchain-community
langchain-core
langchain-experimental
transformers
torch
beautifulsoup4
requests
selenium
pandas
numpy
scikit-learn
matplotlib
seaborn
python-dotenv
chromadb
sentence-transformers
webdriver-manager
lxml
fake-useragent